+++ src/constants.ts
@@ -2,3 +2,5 @@
 export const WIDTH = 2000;
 export const HEIGHT = 1080;
 export const MARGIN = 25;
+
+export const TEAM_HEIGHT = 100;
